Move FormState type higher up
