+++ apps/back-office/src/types.ts
@@ -1,0 +1,5 @@
+export type FormState = {
+  formData?: FormData
+  systemError?: unknown
+  validationErrors?: { key: string; message: string }[]
+}
